feat: fully implement service features (#20)
* fix: optimise select users by service

* feat: add endpoint for getusersbyservice

* feat: add bulk user service change to endpoints

* feat: add service page

* fix: fix code smells

* feat: update navbar headers

* feat: add create services gui

* fix: clean up

* fix: hopefully fix build process

* feat: add delete service action

* fix: move to node to prevent segfaults

* feat: finalise services page

* chore: clean up code smells

* chore: revert breaking changes, optimise further

* feat: add minio

* feat: add fullstack support for minio

* fix: auto test failure

* fix: fix incorrect placeholder image pathing

* fix: simplify initialisation

* fix: revert accidental removal of build test

* chore: update readme

* chore: standardise import aliases

* fix: incorrect aliases

* fix: show error on invalid user permission combination

* feat: add endpoint for getting service sessions

* feat: allow optional specification of service_id when getting all

* feat: tweak behaviour of /service/session/get_all

* feat: add service session display

* chore: standardise import aliases for service sessions

* fix: fix build

* feat: allow for many roles to access same endpoint

* chore: clean up providers

* feat: add delete service session use bulk endpoint

* feat: add edit sessions capability

* fix: fix smells

* fix: end time before start time

* fix: standardise naming

* feat: add service sessions

* feat: delete action

* fix: misc fixes

* chore: clean up code

* refactor: use resusable CRUDModal for add and delete service

* fix: prevent already selected ICs from being selected again

* refactor: split types into file

* fix: incorrect alignment for svc details

* chore: allow update cascade

* chore: prettier backend

* refactor: organise code

* feat: add edit service

* fix: incorrect import of service type

* fix: restrict service ic options to valid options

* fix: incorrect dumping of promotional_image in db

* fix: invalid image url caused by invalid file

* refactor: remove stateful service in servicebox

* fix: build errs

* fix: smells

* fix: fix more code smells

* feat: add scheduler to dump hashes

* feat: add get active service sessions

* fix: incorrect mapping for active sessions get

* fix: incorrect key type for attendance status

* feat: UI improvements for service session users display

* feat: add attendance page

* feat: add QR functionality

* feat: add support for verify attendance

* feat: add frontend verify attendance functionality

* chore: backend prettier

* feat: finalise design

* fix: code smells

* fix: more adjustments

* feat: add support for specifying user when getting endpoint

* fix: misc cleanup and fixes

* fix: crash when user has no permissions

* fix: check for corrupt user type onload

* feat: add support for profile pictures

* fix: security issues with endpoints

* chore: prettier

* fix: crash in user integrity check

* feat: overview page

* feat: add support for updating profile picture

* feat: Add getAllServiceSessionsByUser method to UserModel

* fix: add additional metadata in query

* fix: build error

* feat: add frontend support for service sessions tab

* fix: fix code smell

* feat: happy new year!

* fix: non descriptive error when no active sessions found

* feat: design service sessionn card

* fix: clean up code

* fix: sort in order

* chore: service box styling change

* feat: absence page

* fix: update tests to reflect sort changes

* feat: add support for querying for ad hoc sessions

* fix: consistant api responses

* fix: crash when no one is IC

* feat: add service card to profile

* fix: code smells

+++ interapp-backend/api/routes/endpoints/announcement.ts
@@ -1,6 +1,6 @@
 import { Router } from 'express';
 import { HTTPError, HTTPErrorCode } from '@utils/errors';
-import { validateRequiredFields, verifyJWT, verifyRequiredRole } from '../middleware';
+import { validateRequiredFields, verifyJWT, verifyRequiredPermission } from '../middleware';
 import { AnnouncementModel } from '@models/announcement';
 import { Permissions } from '@utils/permissions';
 
@@ -10,7 +10,7 @@ announcementRouter.post(
   '/',
   validateRequiredFields(['creation_date', 'title', 'description', 'username'], ['attatchment']),
   verifyJWT,
-  verifyRequiredRole(Permissions.EXCO),
+  verifyRequiredPermission(Permissions.EXCO),
   async (req, res) => {
     const ISO8601Regex = /\d{4}-[01]\d-[0-3]\dT[0-2]\d:[0-5]\d([+-][0-2]\d:[0-5]\d|Z)/;
     if (!ISO8601Regex.test(req.body.creation_date)) {
@@ -49,7 +49,7 @@ announcementRouter.patch(
     ['creation_date', 'username', 'title', 'description', 'attatchment'],
   ),
   verifyJWT,
-  verifyRequiredRole(Permissions.EXCO),
+  verifyRequiredPermission(Permissions.EXCO),
   async (req, res) => {
     const ISO8601Regex = /\d{4}-[01]\d-[0-3]\dT[0-2]\d:[0-5]\d([+-][0-2]\d:[0-5]\d|Z)/;
     if (req.body.creation_date && !ISO8601Regex.test(req.body.creation_date)) {
@@ -69,7 +69,7 @@ announcementRouter.delete(
   '/',
   validateRequiredFields(['announcement_id']),
   verifyJWT,
-  verifyRequiredRole(Permissions.EXCO),
+  verifyRequiredPermission(Permissions.EXCO),
   async (req, res) => {
     await AnnouncementModel.deleteAnnouncement(Number(req.body.announcement_id));
     res.status(204).send();
